Feat: FeedSection 서버/클라이언트 분리 및 교회소식 API 연동
- FeedSection(서버)이 getNotices를 직접 호출하도록 변경
- 탭 인터랙션은 FeedContent(클라이언트)로 분리
- 아이템 구조를 제목 행 + 메타데이터 행으로 재편
- N 뱃지 정사각형 처리, 타이틀 ellipsis 및 레이아웃 overflow 수정
