+++ src/app/_component/home/FeedSection.module.scss
@@ -1,10 +1,3 @@
-// 로컬 변수 (디자인 토큰 미포함)
-$color-feed-primary: #5c6b4f;
-$color-feed-primary-light: #eef1eb;
-$color-feed-accent: #c49352;
-$radius-list: 1.4rem;
-$radius-list-lg: 1.6rem;
-
 // ─── Section ────────────────────────────────────────
 .section {
   padding: $padding-section-y 0;
@@ -36,235 +29,5 @@ $radius-list-lg: 1.6rem;
     margin: $spacing-xxs 0;
     display: block;
     font-size: $font-size-lg !important;
-  }
-}
-
-// ─── Tab Switcher (Mobile Only) ─────────────────────
-.tab_switcher {
-  display: flex;
-  gap: $spacing-xxs;
-  margin-bottom: $spacing-md;
-
-  @include respond-up($breakpoint-lg) {
-    display: none;
-  }
-}
-
-.tab_active,
-.tab_inactive {
-  flex: 1;
-  padding: $spacing-sm 0;
-  border-radius: $radius-card;
-  border: none;
-  font-size: $font-size-default;
-  font-weight: $font-weight-semibold;
-  font-family: $font-family-base;
-  cursor: pointer;
-  transition:
-    background $transition-duration-normal $transition-easing-default,
-    color $transition-duration-normal $transition-easing-default;
-}
-
-.tab_active {
-  background: $color-gray-900;
-  color: $color-white;
-}
-
-.tab_inactive {
-  background: $color-cream;
-  color: $color-text-secondary;
-}
-
-// ─── Grid ───────────────────────────────────────────
-.grid {
-  @include respond-up($breakpoint-lg) {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: $spacing-xxl;
-  }
-}
-
-// ─── Column ─────────────────────────────────────────
-.column_hidden {
-  display: none;
-
-  @include respond-up($breakpoint-lg) {
-    display: block;
-  }
-}
-
-// ─── Column Header (PC Only) ────────────────────────
-.column_header {
-  display: none;
-
-  @include respond-up($breakpoint-lg) {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: $spacing-sm;
-  }
-}
-
-.column_title {
-  @include text-card-title;
-  display: flex;
-  align-items: center;
-  gap: $spacing-xs;
-}
-
-.color_bar_news,
-.color_bar_sharing {
-  width: $spacing-xxs;
-  height: 2.2rem;
-  border-radius: $spacing-nano;
-  display: inline-block;
-  flex-shrink: 0;
-  transform: translateY(-0.1rem);
-}
-
-.color_bar_news {
-  background: $color-feed-primary;
-}
-
-.color_bar_sharing {
-  background: $color-feed-accent;
-}
-
-.more_link {
-  @include text-caption;
-  font-weight: $font-weight-semibold;
-  text-decoration: none;
-  color: $color-feed-accent;
-  transition: opacity $transition-duration-normal $transition-easing-default;
-
-  &:hover {
-    opacity: 0.7;
-  }
-}
-
-// ─── List Card ──────────────────────────────────────
-.list_card {
-  background: $color-white;
-  border-radius: $radius-list;
-  border: 1px solid $color-border-default;
-  overflow: hidden;
-
-  @include respond-up($breakpoint-lg) {
-    border-radius: $radius-list-lg;
-  }
-}
-
-// ─── Item Row ───────────────────────────────────────
-.item {
-  padding: $padding-card-compact;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-
-  &:not(:last-child) {
-    border-bottom: 1px solid $color-gray-200;
-  }
-
-  @include respond-up($breakpoint-lg) {
-    padding: $padding-card;
-    transition: background $transition-duration-normal $transition-easing-default;
-
-    &:hover {
-      background: $color-cream;
-    }
-  }
-}
-
-// ─── Badges ─────────────────────────────────────────
-.badge_bulletin {
-  color: $color-feed-primary;
-  font-size: $font-size-xs;
-  font-weight: $font-weight-semibold;
-  background: $color-feed-primary-light;
-  padding: $padding-inline-xs;
-  border-radius: $radius-control;
-  margin-right: $spacing-xs;
-  flex-shrink: 0;
-
-  @include respond-up($breakpoint-lg) {
-    margin-right: $spacing-sm;
-  }
-}
-
-.badge_new {
-  color: $color-white;
-  font-size: $font-size-xs;
-  font-weight: $font-weight-bold;
-  background: $color-status-danger;
-  border-radius: $radius-control;
-  padding: $spacing-min $spacing-xxs;
-  margin-right: $spacing-xxs;
-  flex-shrink: 0;
-
-  @include respond-up($breakpoint-lg) {
-    margin-right: $spacing-xs;
-  }
-}
-
-// ─── Item Content ───────────────────────────────────
-.item_title {
-  color: $color-text-default;
-  font-size: $font-size-default;
-  font-weight: $font-weight-medium;
-  flex: 1;
-  min-width: 0;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-
-  @include respond-up($breakpoint-lg) {
-    font-size: $font-size-lg;
-  }
-}
-
-.item_text {
-  flex: 1;
-  min-width: 0;
-  display: flex;
-  flex-direction: column;
-
-  .item_title {
-    flex: none;
-  }
-}
-
-.item_author {
-  @include text-caption;
-  margin-top: $spacing-min;
-
-  @include respond-up($breakpoint-lg) {
-    margin-top: $spacing-nano;
-  }
-}
-
-.item_date {
-  @include text-caption;
-  flex-shrink: 0;
-}
-
-// ─── More Button (Mobile Only) ──────────────────────
-.more_button {
-  display: block;
-  width: 100%;
-  padding: $spacing-xs 0;
-  border-radius: $radius-list;
-  border: 1px solid $color-border-default;
-  background: transparent;
-  color: $color-text-secondary;
-  font-size: $font-size-sm;
-  font-weight: $font-weight-semibold;
-  font-family: $font-family-base;
-  cursor: pointer;
-  margin-top: $spacing-sm;
-  text-align: center;
-  text-decoration: none;
-
-  @include respond-up($breakpoint-lg) {
-    display: none;
   }
 }
