+++ src/app/page.tsx
@@ -1,11 +1,4 @@
-import {
-  Banner,
-  QuickAccess,
-  RecentSermons,
-  FeedSection,
-  AboutOurChurch,
-  ChurchVision
-} from './_component/home';
+import { Banner, QuickAccess, RecentSermons, FeedSection } from './_component/home';
 
 export default async function Home() {
   return (
